Solidify ghost if something is typed outside it
Ghosts should always be auto-expanded to the far end of the block.
Ideally you wouldn't even be able to move the cursor outside a ghost
paren, but that's impossible to do cleanly in the current architecture
(we can't just override .moveOutOf(), the cursor could end up there by
selecting the one-sided paren group and then using an arrow key to land
on either side of the, or by clicking).

However, we can still maintain the invariant that that ghosts are always
auto-expanded to the far end of the block by, whenever something is
inserted between the ghost and the end, solidfying the ghost and making
the paren pair not one-sided anymore.