Pair up opposing one-sided brackets in more cases
Specifically, the 2 cases that arise from trying to change what's
wrapped in parens by typing before deleting. For example, say we have:

  (1+2)+3+4

and we want to change that to:

  (1+2+3)+4

We can delete the close-paren, which becomes a ghost and auto-expands to
the far end, and then type a close-paren where we want to close the
paren pair. But if we instead type the close-paren first, we get:

  ((1+2)+3)+4

where the outer open-paren is a ghost. Then what if we delete the inner
close-paren? Before this commit, the inner close-paren just becomes a
ghost and auto-expands to the far end of its containing block:

  ((1+2+3))+4

where both the outer open-paren and the inner close-paren are now
ghosts, which looks clearly wrong, they clearly look like they should be
paired up.

The other case is analogous but comes from trying to shrink rather than
expand a paren pair, e.g. from:

  (1+2+3)

to:

  (1+2)+3

by way of:

  ((1+2)+3)

where the inner open-paren is a ghost.

See comments on new if-checks. Also, factor out Bracket::closeOpposing()
from Bracket::createLeftOf() and Bracket::unwrap() from
Bracket::deleteSide(), for use when dealing with these new cases.

Recommend -w to view diff.