selectTowards shan't call cursor.startSelection()
Should've been removed from MathCommand::selectTowards way back:
https://github.com/mathquill/mathquill/commit/44e5c97ece0c459fe6714d83e9b27b9015706709#commitcomment-5768270

Should never have been in TextPiece::selectTowards in the first place,
this is what I really meant to do to fix the anticursor's tree pointers.
Doing cursor.startSelection() worked because this line is only reached
in the case that the cursor and anticursor are at the same point in the
tree.