Persistent anticursor ancestor map inst. of Fragment.between
While selection is in progress, the tree can't change (typing calls
Cursor::endSelection via the `notify` system), so there's no reason for
Fragment.between to do its stupidly complicated and prematurely
optimized tree traversal every time, when it's much simpler to get the
whole ancestor chain of the anticursor once, then have the cursor just
check against that to find the LCA and the ends of the selection.

Since this is now inlined directly into Cursor:;select, the test cases
were moved out of tree.test.js into their own select.test.js.

Also, prep for MathCommand::unselectInto using the anticursor's reverse
ancestor chain instead of .selectedOutOf (which is only set if the
selection was initially made by keyboard rather than mouse), and for it
being possible for the directionality of the selection (whether the
cursor was to the left or right of the anticursor, and hence which end
of the selection it should be at) to be available to Cursor::select (see
next commit).