Fix auto-expanded paren block nonempty but gray
Specifically, if you have `()+stuff` (where there should be a gray box
in the empty paren block) and backspace the close-paren, the paren
auto-expands to the right end of the block but the paren block (now
containing `+stuff`) remains gray.

An alternative fix for this is for Cursor::insDirOf (which is called the
line after the next after the one-liner fix) to call this.parent.focus()
instead of directly doing `this.parent.jQ.addClass('hasCursor')`. The
reason it doesn't already is that `.jQ.removeClass('empty')` is normally
unnecessary when inserting the cursor next to something, because if
there's something to insert the cursor next to, the parent must not be
empty. This is the first case ever that might possibly be an exception,
and if we've been micro-optimizing, why stop now? Bad habits die hard.