Fix UnItalicized LaTeX output to use \operatorname{...}
...when outputting UnItalicized's that aren't built-in LaTeX commands,
like arcsec. Also parse \operatorname{...} stupidly, by pretending the
letters inside are just variables and auto-unitalicizing as appropriate.

I was originally going to have 'inj' and 'proj' auto-unitalicize as
\operatorname{inj} and \operatorname{proj}, too, but it was a ton of
code (the cleanup in the previous commit was pretty important in making
the diff readable, though it's helpful even now) and I eventually
realized this was much simpler and I should err on the side of fewer
3-letter auto-unitalicized commands anway. So now, 'inj' and 'proj' no
longer auto-unitalicize at all, only 'injlim' and 'projlim' as a whole
get auto-unitalicized. This is implemented as a special case, but
whatever.

This problem was brought up in #286. Also fixes #289 (including [1] by
only setting a 'last' CSS class but no 'first' class in the middle of
two-word operator names; choice of that over 'first' but no 'last' was
largely arbitrary, mainly it just looked more right to see extra space
after rather than before a word).

[1]: https://github.com/mathquill/mathquill/issues/289#issuecomment-41447574