Edit "prevent focus on mousedown in IE≤8" comment
...to link to my thorough explanation/test case on jsbin, rather than
the jQuery ticket where I merely mention it (it links to that jQuery
ticket anyway).

The fix to prevent focus on mousedown in IE≤8 was committed on accident:
https://github.com/mathquill/mathquill/commit/a4ae77cc5378ed9d25e5c111426a2b55227808e8#diff-1
but it's pretty much exactly what I want so I'm not gonna bother
reverting and then re-committing it. Instead, here I make a minor edit
to the comment, and mention that it fixes #206.