Rename external .mathquill-{textbox => text-field}
Styling should be only done with internal CSS classes, so now the README
refers to .mathquill-text-field and never .mathquill-textbox (which
should be .mq- not .mathquill-).

The new .mathquill-text-field CSS class is entirely external, its only
mention internally is on document ready, find them all and render them
as MathQuill.TextField().

Recommend --word-diff-regex='[a-z-]+|.' to view diff.