Fix .cmd() ReferenceError; rm Cursor::insertCmd
Cursor::insertCmd is used by LatexCommandInput::renderCommand and by
EditableField::cmd (the public API). MathQuill-basic doesn't have
TextBlock's, so passing .cmd() an unrecognized LaTeX command actually
causes a ReferenceError right now, but that is the only place that still
references TextBlock, and only in that edge case, so we still def don't
want TextBlock in MathQuill-basic.

So what I do is say, those uses of Cursor::insertCmd should never have
been unified, inline it as-is into LatexCommandInput::renderCommand and
inline the non-TextBlock part into EditableField::cmd (instead, when
passed an unrecognized backslash command, noop, which is at least
consistent-ish with .latex() (which empties the MathQuill if unable to
parse the LaTeX), though proper error reporting would be nice).

I actually really don't like the unknown-backslash-commands-are-turned-
into-TextBlock's behavior (like \asdf turns into \text{asdf}, which I'm
leaving in LatexCommandInput::renderCommand), that's weird and I want to
get rid of it eventually, but what I want to do instead is force the
cursor to stay in the LatexCommandInput and I haven't really thought
through how to do that and it shouldn't matter for now.

Also, unit test all the new behavior, normal and error conditions; there
were basically no unit tests for the old behavior.