Fix LiveFraction referencing TextBlock in MQ-basic
...and throwing a ReferenceError just for typing '/'.

Instead, `instanceof TextBlock` should just silently fail if there's no
TextBlock, because then a node couldn't possibly be an instance of it.

(I was amazed to learn that `instanceof undefined` errors, I totally
thought it fails silently, apparently instanceof anything that's not a
function errors, so just `|| noop`.)

Also added a MathQuill-basic section to the LiveFraction test.

Recommend -w to view diff.