Fix autocommands' auto-unitalicized commands check
The point of this check to is to prevent an infinite loop from setting
an autocommand like `sin`, which would try to create `\sin`, which is an
UnItalicized that, in createLeftOf, writes the 3 Letter's 's', 'i', and
'n', which if set as an autocommand would try to create `\sin` and hence
infinite loop.

Since this check was first added (#265), it's become possible to
override `UnItalicizedCmds` (#301), which would break the
`UnItalicizedCmds.hasOwnProperty()` check. Instead, now test directly
for whether the autocommand would create an UnItalicized.