Move Controller linking-up into Controller::init #refactor
Factor out the Controller/API object linking-up-to-each-other code from
AbstractMathQuill::initRoot and InnerMathField::init into
Controller::init.

And, InnerMathField::init now does setup first, then puts cursor at the
end after all the math rendered from LaTeX, just like MathField.

It sucks that the completely unrelated `leftRightIntoCmdGoes` option
processing code had to be tweaked, and only because we haven't had a
good architecture for it and opted[1] to override Controller::init,
which is brittle in exactly this case. Luckily that's gonna be fixed a
few commits later on this branch. Also, so glad for the test case[2] or
I would never have noticed.

[1]: https://github.com/mathquill/mathquill/commit/7d2214eb71efbee5b6453791297429668841d7a4#diff-2
[2]: https://github.com/mathquill/mathquill/commit/7d2214eb71efbee5b6453791297429668841d7a4#diff-a3de19b1ac6167152005fb40ac271d47R139