Reduce "unitalicize"/"un-italicize" inconsistency #cosmetic
We weren't being consistent with whether to camelCase it as UnItalicized
and autoUnItalicized or AutoUnitalicized, and whether to dasherize it as
unitalicize and auto-unitalicized or un-italicized. The main problem is
"auto-un-italicize" is weird and awkward, but "unitalicize" almost looks
cognate with "unital", which is a real math word.

From now on, no more "unitalicize", only "un-italicize" and only when
it's absolutely necessary to use the verb; where possible, use "operator
name" and "non-italicized". ("non-italicized" appears in the Wikipedia
article on "Italic type", whereas none of "un-italicized",
"unitalicized", nor "de-italicized" appear; and "non-italicized" has
many more Google search results than any of the others.)

In particulary, externally (in the README), I now call the option
"autoOperatorNames", and talk about "operator names that automatically
become non-italicized", avoiding the verb form completely. (Also clarify
"without typing a backslash first" instead of "outside a
`LatexCommandInput`".)

Instead of the UnItalicized command, it's now the OperatorName command,
especially fitting given the associated generic backslash command is
\operatorname.

Instead of the .mq-un-italicized CSS class, it's now .mq-operator-name.

Instead of "auto-unitalicized command" in comments (and the one error
message), say "operator name". In comments that have to use the verb,
just live with "auto-un-italicize", at least it's consistent with the
.autoUnItalicize() method.

Instead of MAX_UNITALICIZED_LEN, it's now MAX_AUTO_OP_NAME_LEN, and
instead of UnItalicizedCmds, it's AutoOpNames, which matches variable
name AutoCmds better.

To contrast with AutoOpNames, it's now BuiltInOpNames instead of
OperatorNames, and for consistency, it's TwoWordOpNames instead of
TwoWordOps.

Similarly, rename similar-sounding `trigOps` and `trigs` variables to
contrasting names `builtInTrigs` and `autoTrigs`, respectively. (Also,
simplify builtInTrigs, why weren't they all in one string? [1])

[1]: https://github.com/mathquill/mathquill/commit/a4ae77cc5378ed9d25e5c111426a2b55227808e8#commitcomment-7294720