MathQuill.config() instead of MathQuill.options
No more global MathQuill.options default options [47e3895][], which was:
- inflexible: special processing for leftRightIntoCmdGoes was done by
  overriding Controller::init, how ugly is that?
  Was specifically problematic just a few commits ago [6a043a9][].
  The special processing required for autoCommands and autoOperatorNames
  would only be worse.
- a needlessly different syntax from per-field options, in a way that's
  annoying if you set many options, because you'd have to do:

    MathQuill.options.spaceBehavesLikeTab = true;
    MathQuill.options.supSubsRequireOperand = true;
    MathQuill.options.leftRightIntoCmdGoes = 'up';

  Now instead it's

    MathQuill.config({
      spaceBehavesLikeTab: true,
      supSubsRequireOperand: true,
      leftRightIntoCmdGoes: 'up'
    });

Also add to AbstractMathQuill a .config({ ... }) method to dynamically
change the options from what was supplied to the constructor, again with
the same syntax. It's actually not coincidence that these 3 ways of
setting options (global defaults, per-field on creation, per-field
dynamically) all use the same options object format, they in fact share
the same code, so that's cool.

Mention the new .config() methods in the README (the options example now
stores the created math field in a variable so we can talk about
dynamically changing its options).

[47e3895]: 47e38957d14f878a9ddaa921e82ed4cf46ed2d4b
[6a043a9]: 6a043a93cb4e561d54c335857b182dadb9342147