Unit test page complains about uncaught errors
The next commit breaks MathQuill.addAutoCommands(), but that call is
directly in the 'auto-cmds' suite() of typing.test.js rather than in a
test(), so Mocha doesn't notice and in fact ends up not registering any
subsequent tests and may even show "100%" in the top-right corner, with
no indication anywhere on the page of an uncaught exception. Until now.