`autoCommands` option instead of addAutoCommands
Replace MathQuill.addAutoCommands() with `autoCommands` config option
(i.e. for use with either `MathQuill.config({ ... })` or in the options
object passed to the MathQuill.MathField() constructor).

By virtue of the config options architecture, it can now be overridden,
it can be a different set for different math fields, and it can be
changed dynamically on a per-field basis.