Get rid of in-supsub-only fraction styling #simplify
This shit dates back to the very first commit:
https://github.com/mathquill/mathquill/commit/32004248abdf57e0f36f6187963b43112e860f13#diff-ab9b481f2089bb211b4eb2b6612d6159R113

The numerator no longer has any padding anyway, so that rule doesn't do
anything.

The denominator without padding looks terrible even in a SupSub, the
padding I set in the previous commit looks great even in a SupSub, in so
many browsers.