Tweak paren vertical-align to look good in more font-sizes
@eluberoff first noticed that at font-size:20px (that of expressions in
the Desmos Graphing Calculator), parentheses are too low.

Add a test for vertical-align of parens at font-sizes 10-24px, and
optimized for their appearance in latest Chrome (39), Firefox (34), and
IE (11).

Switched from vertical-align:bottom to top so that the transform-origin
can be a distance from the top (I guess the glyph has some padding).

Increased height of parens to always be 1.2x the height of their
contents rather than only 1.05x, so that parens are always taller than
their contents, whereas before when the contents became too big (like
the 1-8 stacked fraction) the parens would be slightly shorter than
their contents.