Replace zero width joiner with zero width space.
`&#8203` is the zero width space character. We replaced most other uses of the
zero width joiner with a zero width space because of a font rendering bug in
Chrome. For whatever reason, this particular example doesn't visually exhibit
the bug, but I think it makes sense to replace zwj here for consistency, and
because a zero width space makes more semantic sense.